Use MUnit in SwaggerSupportSuite
